ConDec-377: Move Calculation of the Report metrics in a seperate class
